+++ libretro-common/include/streams/interface_stream.h
@@ -68,29 +68,29 @@ bool intfstream_resize(intfstream_internal_t *intf,
 bool intfstream_open(intfstream_internal_t *intf,
       const char *path, unsigned mode, unsigned hints);
 
-ssize_t intfstream_read(intfstream_internal_t *intf,
-      void *s, size_t len);
+int64_t intfstream_read(intfstream_internal_t *intf,
+      void *s, uint64_t len);
 
-ssize_t intfstream_write(intfstream_internal_t *intf,
-      const void *s, size_t len);
+int64_t intfstream_write(intfstream_internal_t *intf,
+      const void *s, uint64_t len);
 
 char *intfstream_gets(intfstream_internal_t *intf,
       char *buffer, size_t len);
 
 int intfstream_getc(intfstream_internal_t *intf);
 
-int intfstream_seek(intfstream_internal_t *intf,
-      int offset, int whence);
+int64_t intfstream_seek(intfstream_internal_t *intf,
+      int64_t offset, int whence);
 
 void intfstream_rewind(intfstream_internal_t *intf);
 
-int intfstream_tell(intfstream_internal_t *intf);
+int64_t intfstream_tell(intfstream_internal_t *intf);
 
 void intfstream_putc(intfstream_internal_t *intf, int c);
 
 int intfstream_close(intfstream_internal_t *intf);
 
-ssize_t intfstream_get_size(intfstream_internal_t *intf);
+int64_t intfstream_get_size(intfstream_internal_t *intf);
 
 int intfstream_flush(intfstream_internal_t *intf);
